# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-spherevolume",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the volume of a sphere `V = (4/3) * pi * r^3`.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits float primitives, exactly as specified:\n\n1. ask `srvcs-pi` (a constant service, called with an empty body) for `p`;\n2. `r2 = floatmultiply(radius, radius)`;\n3. `r3 = floatmultiply(r2, radius)`;\n4. `t  = floatmultiply(p, r3)`;\n5. `t2 = floatmultiply(4, t)`;\n6. `result = floatdivide(t2, 3)`.\n\nValidation is not handled here: this service never calls `srvcs-isnumber`.\nIts dependencies validate their own operands, and any `422` they raise is\nforwarded verbatim. If a dependency is unreachable it reports itself\ndegraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SphereVolumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "radius"
        ],
        "properties": {
          "radius": {
            "type": "object",
            "description": "The radius of the sphere."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "SphereVolumeResponse": {
        "type": "object",
        "required": [
          "radius",
          "result"
        ],
        "properties": {
          "radius": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}